+++ lib/app_info.dart
@@ -1,6 +1,6 @@
 class AppInfo {
   static const String name = 'Narrative Data App';
-  static const String version = '1.119.0';
+  static const String version = '1.120.0';
 
   static String get displayName => '$name v$version';
 }